Moved LoginButton and SignOutButton into new AuthButtons.tsx file
This has been done to keep the file structure of the app intuitive.
