+++ client/src/components/Common/CustomButtons.tsx
@@ -29,3 +29,4 @@ const styles = StyleSheet.create({
         marginLeft: Dimensions.get('window').width * 0.01,
     },
 })
+
