+++ client/src/configs/firebaseConfig.ts
